+++ src/components/login/LoginPage2.tsx
@@ -1,0 +1,3 @@
+export default function LoginPage2() {
+  return <div>LoginPage2</div>;
+}
