+++ src/components/login/LoginPage3.tsx
@@ -1,0 +1,3 @@
+export default function LoginPage3() {
+  return <div>Login Page 3</div>;
+}
