+++ src/components/login/LoginPage4.tsx
@@ -1,0 +1,3 @@
+export default function LoginPage4() {
+  return <div>LoginPage4</div>;
+}
