+++ src/components/login/index.ts
@@ -1,0 +1,21 @@
+import LoginModal from './LoginModal';
+import LoginPage0 from './LoginPage0';
+import LoginPage1 from './LoginPage1';
+import LoginPage2 from './LoginPage2';
+import LoginPage3 from './LoginPage3';
+import LoginPage4 from './LoginPage4';
+
+export {
+  LoginModal,
+  LoginPage0,
+  LoginPage1,
+  LoginPage2,
+  LoginPage3,
+  LoginPage4,
+};
+
+// 페이지 0 : 로그인
+// 페이지 1 : 포지션 선택
+// 페이지 2 : 기술스택 선택
+// 페이지 3 : 프로필 사진, 닉네임 변경
+// 페이지 4 : 환영합니다!
